fragment_pokemon_detail.xml redesign

+++ app/src/main/java/com/example/labepamproject/data/network/dto/StatsDTO.kt
@@ -3,7 +3,7 @@ package com.example.labepamproject.data.network
 import com.squareup.moshi.Json
 
 data class PokemonStatsData(
-    @field:Json(name = "base_stat") val base_stat: Int,
+    @field:Json(name = "base_stat") val base_stat: Float,
     val effort: Int,
     val stat: PokemonStatsDetailsData,
 )
